# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: КадастрПро API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /engineers:
    get:
      operationId: listEngineers
      tags: [engineers]
      summary: List engineers
      parameters:
        - name: region
          in: query
          schema:
            type: string
        - name: specialization
          in: query
          schema:
            type: string
        - name: minRating
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: district
          in: query
          schema:
            type: string
        - name: sro
          in: query
          schema:
            type: string
        - name: verifiedOnly
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EngineerList"
...
  /engineers/me/leads:
    get:
      operationId: getMyLeads
      tags: [engineers]
      summary: Get leads for the authenticated engineer
      parameters:
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadList"
...
  /engineers/me:
    get:
      operationId: getMyEngineerProfile
      tags: [engineers]
      summary: Get current engineer profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Engineer"
    put:
      operationId: updateMyEngineerProfile
      tags: [engineers]
      summary: Update current engineer profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EngineerUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Engineer"
  /engineers/verify:
    post:
      operationId: verifyEngineer
      tags: [engineers]
      summary: Verify engineer registry number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResult"
  /engineers/{engineerId}:
    get:
      operationId: getEngineer
      tags: [engineers]
      summary: Get engineer by ID
      parameters:
        - name: engineerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Engineer"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
        telegram:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isBlocked:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, email, role, createdAt]
...
    PortfolioItem:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
        region:
          type: string
        year:
          type: integer
        area:
          type: ["string", "null"]
      required: [title, description, type, region, year]
...
    Engineer:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        registryNumber:
          type: string
        specializations:
          type: array
          items:
            type: string
        region:
          type: string
        regions:
          type: array
          items:
            type: string
        district:
          type: ["string", "null"]
        sro:
          type: ["string", "null"]
        experience:
          type: integer
        bio:
          type: ["string", "null"]
        isVerified:
          type: boolean
        isOnline:
          type: boolean
        isHidden:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
        completedOrders:
          type: integer
        responseTime:
          type: ["string", "null"]
        priceFrom:
          type: ["integer", "null"]
        portfolioItems:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioItem"
        isPro:
          type: boolean
        proExpiresAt:
          type: ["string", "null"]
        debtAmount:
          type: integer
        createdAt:
          type: string
        attestatNumber:
          type: ["string", "null"]
        rosreestrStatus:
          type: ["string", "null"]
        sroName:
          type: ["string", "null"]
        rosreestrCheckedAt:
          type: ["string", "null"]
        rosreestrWorksCount:
          type: ["integer", "null"]
        rosreestrRejectionsCount:
          type: ["integer", "null"]
        rosreestrSuspensionsCount:
          type: ["integer", "null"]
        rosreestrRejectionRate:
          type: ["number", "null"]
      required: [id, userId, user, registryNumber, specializations, region, regions, experience, isVerified, isOnline, isHidden, rating, reviewCount, completedOrders, portfolioItems, isPro, debtAmount, createdAt]
    EngineerList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Engineer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [items, total, page, limit]
    EngineerUpdate:
      type: object
      properties:
        specializations:
          type: array
          items:
            type: string
        region:
          type: string
        regions:
          type: array
          items:
            type: string
        experience:
          type: integer
        bio:
          type: string
        phone:
          type: string
        telegram:
          type: string
        whatsapp:
          type: string
        responseTime:
          type: string
        priceFrom:
          type: integer
        isOnline:
          type: boolean
        portfolioItems:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioItem"
    VerifyInput:
      type: object
      properties:
        registryNumber:
          type: string
      required: [registryNumber]
    VerifyResult:
      type: object
      properties:
        isValid:
          type: boolean
        engineerName:
          type: ["string", "null"]
        message:
          type: string
      required: [isValid, message]
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        engineerId:
          type: integer
        engineerName:
          type: string
        serviceType:
          type: string
        leadCost:
          type: integer
        paymentStatus:
          type: string
        createdAt:
          type: string
      required: [id, orderId, engineerId, serviceType, leadCost, paymentStatus, createdAt]
    LeadList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [items, total, page, limit]
...